Folding: unifying alphas definitions
There must be only one definition of alphas, up-to-now. We don't need to
duplicate code N times.
