+++ folding/src/examples/generic.rs
@@ -2,7 +2,7 @@ use crate::{
     expressions::{FoldingColumnTrait, FoldingCompatibleExpr, FoldingCompatibleExprInner},
     FoldingConfig, Sponge,
 };
-use ark_ff::{Field, One, Zero};
+use ark_ff::{Field, Zero};
 use kimchi::curve::KimchiCurve;
 use mina_poseidon::{
     constants::PlonkSpongeConstantsKimchi,
@@ -10,11 +10,6 @@ use mina_poseidon::{
     FqSponge,
 };
 use poly_commitment::PolyComm;
-use std::{
-    iter::successors,
-    rc::Rc,
-    sync::atomic::{AtomicUsize, Ordering},
-};
 
 #[cfg(not(test))]
 use log::debug;
@@ -54,58 +49,7 @@ impl FoldingColumnTrait for Column {
     }
 }
 
-// 3. We define the combinators that will be used to fold the constraints,
-// called the "alphas".
-// The alphas are exceptional, their number cannot be known ahead of time as it
-// will be defined by folding.
-// The values will be computed as powers in new instances, but after folding
-// each alpha will be a linear combination of other alphas, instand of a power
-// of other element. This type represents that, allowing to also recognize
-// which case is present.
-#[derive(Debug, Clone)]
-pub enum Alphas {
-    Powers(Fp, Rc<AtomicUsize>),
-    Combinations(Vec<Fp>),
-}
-
-impl Alphas {
-    pub fn new(alpha: Fp) -> Self {
-        Self::Powers(alpha, Rc::new(AtomicUsize::from(0)))
-    }
-    pub fn get(&self, i: usize) -> Option<Fp> {
-        match self {
-            Alphas::Powers(alpha, count) => {
-                let _ = count.fetch_max(i + 1, Ordering::Relaxed);
-                let i = [i as u64];
-                Some(alpha.pow(i))
-            }
-            Alphas::Combinations(alphas) => alphas.get(i).cloned(),
-        }
-    }
-    pub fn powers(self) -> Vec<Fp> {
-        match self {
-            Alphas::Powers(alpha, count) => {
-                let n = count.load(Ordering::Relaxed);
-                let alphas = successors(Some(Fp::one()), |last| Some(*last * alpha));
-                alphas.take(n).collect()
-            }
-            Alphas::Combinations(c) => c,
-        }
-    }
-    pub fn combine(a: Self, b: Self, challenge: Fp) -> Self {
-        let a = a.powers();
-        let b = b.powers();
-        assert_eq!(a.len(), b.len());
-        let comb = a
-            .into_iter()
-            .zip(b)
-            .map(|(a, b)| a + b * challenge)
-            .collect();
-        Self::Combinations(comb)
-    }
-}
-
-// 4. We define different traits that can be used generically by the folding
+// 3. We define different traits that can be used generically by the folding
 // examples.
 // It can be used by "pseudo-provers".
 pub(crate) trait Provide<C: FoldingConfig> {
